+++ R/module-define.R
@@ -605,6 +605,7 @@ setMethod(
 #'
 #' @keywords internal
 #' @importFrom data.table setnames
+#' @importFrom tools file_ext
 #' @rdname fillInputRows
 .fillInputRows <- function(inputDF, startTime) {
   factorCols <- sapply(inputDF, is.factor)
@@ -646,14 +647,14 @@ setMethod(
     if (any(is.na(inputDF2[, "fun"]))) {
       .fileExts <- .fileExtensions()
       fl <- inputDF2$file
-      exts <- na.omit(match(fileExt(fl), .fileExts[, "exts"]))
+      exts <- na.omit(match(file_ext(fl), .fileExts[, "exts"]))
       inputDF2$fun[is.na(inputDF2$fun)] <- .fileExts[exts, "fun"]
     }
 
     if (any(is.na(inputDF2[, "package"]))) {
       .fileExts <- .fileExtensions()
       fl <- inputDF2$file
-      exts <- match(fileExt(fl), .fileExts[, "exts"])
+      exts <- match(file_ext(fl), .fileExts[, "exts"])
       inputDF2$package[is.na(inputDF2$package)]  <- .fileExts[exts, "package"]
     }
     inputDF[!objectsOnly, ] <- inputDF2
@@ -668,6 +669,7 @@ setMethod(
 #'
 #' @keywords internal
 #' @importFrom data.table setnames
+#' @importFrom tools file_ext
 #' @rdname fillOutputRows
 .fillOutputRows <- function(outputDF, endTime) {
   needRenameArgs <- grepl(names(outputDF), pattern = "arg[s]?$")
@@ -699,7 +701,7 @@ setMethod(
   if (any(is.na(outputDF[, "fun"]))) {
     .fileExts <- .saveFileExtensions()
     fl <- outputDF$file
-    exts <- fileExt(fl)
+    exts <- file_ext(fl)
     if (any(is.na(fl)) | any(!nzchar(exts, keepNA = TRUE))) {
       outputDF$fun[is.na(fl) | (!nzchar(exts, keepNA = TRUE))] <- .fileExts$fun[1]
     }
@@ -712,12 +714,12 @@ setMethod(
   if (any(is.na(outputDF[, "package"]))) {
     .fileExts <- .saveFileExtensions()
     fl <- outputDF$file
-    exts <- fileExt(fl)
+    exts <- file_ext(fl)
     if (any(is.na(fl)) | any(!nzchar(exts, keepNA = TRUE))) {
       outputDF$package[is.na(fl) | (!nzchar(exts, keepNA = TRUE))] <- .fileExts$package[1]
     }
     if (any(is.na(outputDF[, "package"]))) {
-      exts <- na.omit(match(fileExt(fl), .fileExts[, "exts"]) )
+      exts <- na.omit(match(file_ext(fl), .fileExts[, "exts"]) )
       outputDF$package[is.na(outputDF$package)] <- .fileExts[exts, "package"]
     }
   }